requirements.txt
requests
beautifulsoup4
pillow
matplotlib
fastapi
uvicorn
pydantic
opencv-python
pytesseract
numpy